[REMOVE] Substitui o test_atividade pelo test_agenda.
